move method of findNearestPossibleLink to Utils
